refactor: move oci referrers code
